Did some styling on the Admin pages.
NOTE: I came home really and was having some issues with the NuGet package manager which wouldn't let me run the local project on my machine. I was unable to find a solution after looking through the internet. So I did the best I could styling the pages without being about to view the physical changes in the web browser. Because of this, the styling won't look as great but hopefully it will look somewhat decent for presentation. I'll fine tune things as soon as I'm able to fix the issue. My sincere apologies.